+++ trunk/src/basemode/baseactions.cpp
@@ -33,12 +33,12 @@ This file is part of Liberal Crime Squad.                                       
 /* base - burn the flag */
 void burnflag(void)
 {
-   int flagparts=112;
-   long flag[16][7][4];
-   int x;
-   int y;
+   int32 flagparts=112;
+   int32 flag[16][7][4];
+   int32 x;
+   int32 y;
 
-   for(int p=0;p<7;p++)
+   for(int32 p=0;p<7;p++)
    {
       if(p<3)
       {
@@ -213,15 +213,15 @@ void orderparty(void)
 {
    if(activesquad==NULL)return;
 
-   int partysize=0;
-   for(int p=0;p<6;p++)
+   int32 partysize=0;
+   for(int32 p=0;p<6;p++)
    {
       if(activesquad->squad[p]!=NULL)partysize++;
    }
 
    if(partysize<=1)return;
 
-   int spot=0;
+   int32 spot=0;
 
    do
    {
@@ -237,7 +237,7 @@ void orderparty(void)
 
       refresh();
 
-      int c=getch();
+      int32 c=getch();
       translategetch(c);
 
       if(c==10)return;
@@ -257,8 +257,8 @@ void orderparty(void)
 /* base - go forth to stop evil */
 void stopevil(void)
 {
- int l = 0;
- int p = 0;
+ int32 l = 0;
+ int32 p = 0;
 
    if(activesquad==NULL)return;
 
@@ -275,10 +275,10 @@ void stopevil(void)
       }
    }
 
-   int page=0;
-   long loc=-1;
+   int32 page=0;
+   int32 loc=-1;
 
-   vector<long> temploc;
+   vector<int32> temploc;
    for(l=0;l<location.size();l++)
    {
       if(location[l]->parent==loc&&location[l]->renting>=0)temploc.push_back(l);
@@ -305,12 +305,12 @@ void stopevil(void)
          addlocationname(location[loc]);
       }
 
-      int y=10;
+      int32 y=10;
       for(p=page*11;p<temploc.size()&&p<page*11+11;p++)
       {
          set_color(COLOR_WHITE,COLOR_BLACK,0);
          move(y,0);
-         addch(y-10+(int)'A');
+         addch(y-10+(int32)'A');
          addstr(" - ");
 
          addlocationname(location[temploc[p]]);
@@ -390,7 +390,7 @@ void stopevil(void)
 
       refresh();
 
-      int c=getch();
+      int32 c=getch();
       translategetch(c);
 
       //PAGE UP
@@ -400,10 +400,10 @@ void stopevil(void)
 
       if(c>='a'&&c<='k')
       {
-         int sq=page*11+(int)(c-'a');
+         int32 sq=page*11+(int32)(c-'a');
          if(sq<temploc.size()&&sq>=0)
          {
-            int oldloc=loc;
+            int32 oldloc=loc;
             loc=temploc[sq];
             temploc.clear();
 
@@ -553,8 +553,8 @@ char liberalagenda(char won)
       case 2:addstr("Elite Liberal");break;
    }
 
-   int y=22-LAWNUM;
-   for(int l=0;l<LAWNUM;l++)
+   int32 y=22-LAWNUM;
+   for(int32 l=0;l<LAWNUM;l++)
    {
       if(won==-1)set_color(COLOR_RED,COLOR_BLACK,1);
       else
@@ -687,12 +687,12 @@ char liberalagenda(char won)
       y++;
    }
 
-   int housemake[5]={0,0,0,0,0};
-   for(int h=0;h<435;h++)
+   int32 housemake[5]={0,0,0,0,0};
+   for(int32 h=0;h<435;h++)
    {
       housemake[house[h]+2]++;
    }
-   int lsum=housemake[3]+housemake[4]
+   int32 lsum=housemake[3]+housemake[4]
       -housemake[0]-housemake[1];
    if(lsum<=-145)set_color(COLOR_RED,COLOR_BLACK,1);
    else if(lsum<0)set_color(COLOR_MAGENTA,COLOR_BLACK,1);
@@ -716,8 +716,8 @@ char liberalagenda(char won)
       addstr(num);addstr("Cons+");
    }
 
-   int senatemake[5]={0,0,0,0,0};
-   for(int s=0;s<100;s++)
+   int32 senatemake[5]={0,0,0,0,0};
+   for(int32 s=0;s<100;s++)
    {
       senatemake[senate[s]+2]++;
    }
@@ -750,8 +750,8 @@ char liberalagenda(char won)
       addstr("The Congress consists of CEOs and televangelists.");
    }
 
-   int elibjudge=0;
-   for(int c=0;c<9;c++)
+   int32 elibjudge=0;
+   for(int32 c=0;c<9;c++)
    {
       if(court[c]==2)elibjudge++;
    }
@@ -778,7 +778,7 @@ char liberalagenda(char won)
    {
       y=0;
 
-      for(int c=0;c<9;c++)
+      for(int32 c=0;c<9;c++)
       {
          if(court[c]==-2)set_color(COLOR_RED,COLOR_BLACK,1);
          else if(court[c]==-1)set_color(COLOR_MAGENTA,COLOR_BLACK,1);
@@ -814,7 +814,7 @@ char liberalagenda(char won)
       do
       {
          refresh();
-         int c=getch();
+         int32 c=getch();
          translategetch(c);
 
          if(c=='l')break;
@@ -831,7 +831,7 @@ char liberalagenda(char won)
       do
       {
          refresh();
-         int c=getch();
+         int32 c=getch();
          translategetch(c);
 
          if(c=='l')break;
@@ -848,7 +848,7 @@ char liberalagenda(char won)
       do
       {
          refresh();
-         int c=getch();
+         int32 c=getch();
          translategetch(c);
 
          if(c=='d')
@@ -869,7 +869,7 @@ char liberalagenda(char won)
 char confirmdisband(void)
 {
    char word[80];
-   int pos=0;
+   int32 pos=0;
 
    switch(LCSrandom(5))
    {
@@ -900,7 +900,7 @@ char confirmdisband(void)
       move(6,0);
       addstr("Type this Liberal phrase to confirm (press a wrong letter to rethink it):");
 
-      for(int x=0;x<strlen(word);x++)
+      for(int32 x=0;x<strlen(word);x++)
       {
          move(8,x);
          if(x==pos)set_color(COLOR_GREEN,COLOR_BLACK,0);
@@ -909,7 +909,7 @@ char confirmdisband(void)
          addch(word[x]);
       }
 
-      int c=getch();
+      int32 c=getch();
       translategetch(c);
 
       if((c==word[pos])||((c+'A'-'a')==word[pos]))
@@ -919,14 +919,14 @@ char confirmdisband(void)
          if(pos>=strlen(word))
          {
             //SET UP THE DISBAND
-            for(int p=0;p<pool.size();p++)
+            for(int32 p=0;p<pool.size();p++)
             {
                if(pool[p]->alive&&
                   !(pool[p]->flag & CREATUREFLAG_SLEEPER))
                {
                   pool[p]->activity.type=ACTIVITY_NONE;
                   pool[p]->clinic=0;
-                  for(int i=0;i<LAWFLAGNUM;i++)
+                  for(int32 i=0;i<LAWFLAGNUM;i++)
                   {
                      pool[p]->lawflag[i]=0;
                   }
@@ -952,10 +952,10 @@ char confirmdisband(void)
 /* base - invest in this location */
 void investlocation(void)
 {
-   int loc=selectedsiege;
-   int ppress=0;
+   int32 loc=selectedsiege;
+   int32 ppress=0;
 
-   for(int i=0;i<location.size();i++)
+   for(int32 i=0;i<location.size();i++)
    {
       if(location[i]->compound_walls==COMPOUND_PRINTINGPRESS)ppress=1;
    }
@@ -1041,7 +1041,7 @@ void investlocation(void)
       move(16,1);
       addstr("Enter - Done");
 
-      int c=getch();
+      int32 c=getch();
       translategetch(c);
 
       if(c==10)break;
@@ -1172,10 +1172,10 @@ void investlocation(void)
 /* base - assign a vehicle to this squad */
 void setvehicles(void)
 {
-   int p, l;
+   int32 p, l;
    if(activesquad==NULL)return;
 
-   int page=0;
+   int32 page=0;
 
    do
    {
@@ -1187,7 +1187,7 @@ void setvehicles(void)
 
       printparty();
 
-      int x=1,y=10;
+      int32 x=1,y=10;
       char str[200],str2[200];
 
       for(l=page*18;l<vehicle.size()&&l<page*18+18;l++)
@@ -1283,12 +1283,12 @@ void setvehicles(void)
 
       refresh();
 
-      int c=getch();
+      int32 c=getch();
       translategetch_cap(c);
 
       if(c>='A'&&c<='R')
       {
-         int slot=c-'A'+page*18;
+         int32 slot=c-'A'+page*18;
 
          if(slot>=0&&slot<vehicle.size())
          {
@@ -1298,7 +1298,7 @@ void setvehicles(void)
 
             refresh();
 
-            int c=getch();
+            int32 c=getch();
             translategetch(c);
 
             if(c>='1'&&c<='6')
@@ -1318,7 +1318,7 @@ void setvehicles(void)
 
       if(c>='a'&&c<='r')
       {
-         int slot=c-'a'+page*18;
+         int32 slot=c-'a'+page*18;
 
          if(slot>=0&&slot<vehicle.size())
          {
@@ -1328,7 +1328,7 @@ void setvehicles(void)
 
             refresh();
 
-            int c=getch();
+            int32 c=getch();
             translategetch(c);
 
             if(c>='1'&&c<='6')
